+++ zboss/include/zcl/zb_zcl_power_config.h
@@ -613,15 +613,16 @@ enum zb_zcl_power_config_battery_alarm_state_e
     @param alarm_mask - pointer to variable to store BatteryAlarmMask attribute
     @param voltage_min_threshold - pointer to variable to store BatteryVoltageMinThreshold attribute
 */
-#define ZB_ZCL_DECLARE_POWER_CONFIG_ATTRIB_LIST(attr_list,                                  \
-    voltage, size, quantity, rated_voltage, alarm_mask, voltage_min_threshold)              \
-  ZB_ZCL_START_DECLARE_ATTRIB_LIST(attr_list)                                               \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_ID, (voltage))              \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_SIZE_ID, (size))                    \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_QUANTITY_ID, (quantity))            \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_RATED_VOLTAGE_ID, (rated_voltage))  \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_ALARM_MASK_ID, (alarm_mask ))       \
-  ZB_ZCL_SET_ATTR_DESC(ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_MIN_THRESHOLD_ID, (voltage_min_threshold))  \
+#define ZB_ZCL_DECLARE_POWER_CONFIG_ATTRIB_LIST(attr_list,                                   \
+                                                voltage, size, quantity, rated_voltage,      \
+                                                alarm_mask, voltage_min_threshold)           \
+  ZB_ZCL_START_DECLARE_ATTRIB_LIST(attr_list)                                                \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_ID(voltage, ),             \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_SIZE_ID(size, ),                   \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_QUANTITY_ID(quantity, ),           \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_RATED_VOLTAGE_ID(rated_voltage, ), \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_ALARM_MASK_ID(alarm_mask, ),       \
+  ZB_SET_ATTR_DESCR_WITH_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_MIN_THRESHOLD_ID(voltage_min_threshold, ), \
   ZB_ZCL_FINISH_DECLARE_ATTRIB_LIST
 
 /** @cond internals_doc */
@@ -706,7 +707,7 @@ enum zb_zcl_power_config_battery_alarm_state_e
     remaining, threshold1, threshold2, threshold3, min_threshold, percent_threshold1,       \
     percent_threshold2, percent_threshold3, alarm_state)                                    \
   ZB_ZCL_START_DECLARE_ATTRIB_LIST(attr_list)                                               \
-  ZB_ZCL_POWER_CONFIG_BATTERY_ATTRIB_LIST_EXT(,                                             \
+  ZB_ZCL_POWER_CONFIG_BATTERY_ATTRIB_LIST_EXT(bat_num,                                      \
     voltage, size, quantity, rated_voltage, alarm_mask, voltage_min_threshold,              \
     remaining, threshold1, threshold2, threshold3, min_threshold, percent_threshold1,       \
     percent_threshold2, percent_threshold3, alarm_state)                                    \